Move MacAddress to separate module.
